requirements/base.txt
Django==1.11
pymysql==0.7.11
redis==2.10.5
jsonfield==1.0.3
djangorestframework==3.4.0
django-filter==1.0.1
